# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Sagara Hackathon App",
        "title": "Sagara Hackathon API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://sagaratechnology.com/",
            "email": "zeptian15@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login/admin": {
            "post": {
                "description": "Login For Superadmin, Admin, HR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Admin",
                "parameters": [
                    {
                        "description": "Body Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/src.AuthSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/src.BaseFailure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "src.AuthSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/src.Token"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "src.BaseFailure": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "$ref": "#/definitions/src.Error"
                },
                "message": {
                    "type": "string",
                    "example": "Failure"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "src.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "src.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}